# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "教室排课、教室资源管理和冲突检测 RESTful API。",
        "title": "教室排课助手 API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/classes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "List classes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Create a class",
                "parameters": [
                    {
                        "description": "class payload",
                        "name": "class",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.CreateClassRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/classes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Get a class",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "class id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Update a class",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "class id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "class payload",
                        "name": "class",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.UpdateClassRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classes"
                ],
                "summary": "Delete a class",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "class id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/classrooms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classrooms"
                ],
                "summary": "List classrooms",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classrooms"
                ],
                "summary": "Create a classroom",
                "parameters": [
                    {
                        "description": "classroom payload",
                        "name": "classroom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.CreateClassroomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/classrooms/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classrooms"
                ],
                "summary": "Get a classroom",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "classroom id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classrooms"
                ],
                "summary": "Update a classroom",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "classroom id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "classroom payload",
                        "name": "classroom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.UpdateClassroomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "classrooms"
                ],
                "summary": "Delete a classroom",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "classroom id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gbschedule_gbschedule_internal_dto.CreateClassRequest": {
            "type": "object",
            "required": [
                "name",
                "student_count"
            ],
            "properties": {
                "grade": {
                    "type": "string",
                    "maxLength": 64
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "student_count": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.CreateClassroomRequest": {
            "type": "object",
            "required": [
                "capacity",
                "code",
                "name"
            ],
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "code": {
                    "type": "string",
                    "maxLength": 64
                },
                "equipment": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.UpdateClassRequest": {
            "type": "object",
            "properties": {
                "grade": {
                    "type": "string",
                    "maxLength": 64
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "student_count": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.UpdateClassroomRequest": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "code": {
                    "type": "string",
                    "maxLength": 64
                },
                "equipment": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
...
    }
}